refactor : nest Dropdown and OptionsInput component under InterviewOption folder
