+++ src/components/Main/InterviewOptionsPanel/InterviewOption/Dropdown/Svg/index.tsx
