+++ src/components/Main/InterviewOptionsPanel/InterviewOption/Dropdown/types/index.ts
@@ -4,5 +4,6 @@ export type DropdownOptionsProps = {
   label: string;
   type: string;
   name: string;
+  tooltipContent: string;
   options?: Options;
 };
