+++ src/components/Main/InterviewOptionsPanel/InterviewOption/index.tsx
@@ -1,15 +1,16 @@
 import { useState, useRef } from 'react';
-import OptionsDropdown from '../Dropdown';
-import OptionsInput from '../Input';
+import OptionLabel from './OptionLabel';
+import OptionsDropdown from './Dropdown';
+import OptionsInput from './Input';
 import useOnClickOutside from '@hooks/useOnClickOutside';
-import { DropdownOptionsProps } from '../Dropdown/types';
+import { DropdownOptionsProps } from './Dropdown/types';
 
 type InterviewOptionProps = {
   option: DropdownOptionsProps;
 };
 
 const InterviewOption = ({ option }: InterviewOptionProps) => {
-  const { label, type, name } = option;
+  const { label, type, name, tooltipContent } = option;
 
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
   const optionRef = useRef(null);
@@ -18,9 +19,7 @@ const InterviewOption = ({ option }: InterviewOptionProps) => {
 
   return (
     <div className="relative w-[200px] mx-4" ref={optionRef}>
-      <label className="block text-sm font-medium leading-6">
-        {label} &#9432;
-      </label>
+      <OptionLabel label={label} tooltipContent={tooltipContent} />
       {type === 'dropdown' && (
         <OptionsDropdown
           option={option}
